feat: New generated content
Update neutron schema resolution

Neutron treats subnetpool.XXX_prefixlen everywhere as integers including
the database (it is even described like that in neutron api-ref itself).
For some unknown reason it returns it as a string though. Maybe because
in the apidef it is defined like `validate: "type:non_negative",
convert_to: convert_to_int`. Modify api_def processing not to look at
validate OR convert_to, but use convert_to results in update of the
schema.


Changes are triggered by https://review.opendev.org/925868

+++ openstack_sdk/src/api/network/v2/network_segment_range/create.rs
@@ -45,11 +45,11 @@ pub struct NetworkSegmentRange<'a> {
 
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default)]
-    pub(crate) maximum: Option<f32>,
+    pub(crate) maximum: Option<i32>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default)]
-    pub(crate) minimum: Option<f32>,
+    pub(crate) minimum: Option<i32>,
 
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default, setter(into))]
