+++ openstack_sdk/src/api/network/v2/qos/alias_minimum_bandwidth_rule/set.rs
@@ -41,7 +41,7 @@ pub struct AliasMinimumBandwidthRule {
 
     #[serde(skip_serializing_if = "Option::is_none")]
     #[builder(default)]
-    pub(crate) min_kbps: Option<f32>,
+    pub(crate) min_kbps: Option<i32>,
 }
 
 #[derive(Builder, Debug, Clone)]
